component refactor
